Moved Home dir related stuff to file-util.cpp/h
