+++ src/drive-dlg.cpp
@@ -5,6 +5,7 @@
  */
 
 #include "drive-dlg.h"
+#include "file-util.h"
 #include "ncwin.h"
 #include "globals.h"
 #include "string-util.h"
@@ -56,21 +57,6 @@ std::string GetDriveTypeString( unsigned int Type )
 	}
 
 	return std::string();
-}
-
-std::vector<unicode_t> GetHomeUriWin()
-{
-	wchar_t homeDrive[0x100];
-	wchar_t homePath[0x100];
-	int l1 = GetEnvironmentVariableW( L"HOMEDRIVE", homeDrive, 0x100 );
-	int l2 = GetEnvironmentVariableW( L"HOMEPATH", homePath, 0x100 );
-
-	if ( l1 > 0 && l1 < 0x100 && l2 > 0 && l2 < 0x100 )
-	{
-		return carray_cat<unicode_t>( Utf16ToUnicode( homeDrive ).data(), Utf16ToUnicode( homePath ).data() );
-	}
-
-	return std::vector<unicode_t>();
 }
 
 #endif // _WIN32
@@ -456,45 +442,4 @@ void SelectDriveDlg( PanelWin* p, PanelWin* OtherPanel )
 		RedoDialog = SelectDriveInternal( p, OtherPanel );
 	}
 	while ( RedoDialog );
-}
-
-void OpenHomeDir( PanelWin* p, PanelWin* OtherPanel )
-{
-#ifdef _WIN32
-	std::vector<unicode_t> homeUri = GetHomeUriWin();
-
-	if ( homeUri.data() )
-	{
-		const std::vector<clPtr<FS>> checkFS =
-		{
-			p->GetFSPtr(),
-			OtherPanel->GetFSPtr()
-		};
-
-		FSPath path;
-		clPtr<FS> fs = ParzeURI( homeUri.data(), path, checkFS );
-
-		if ( fs.IsNull() )
-		{
-			char buf[4096];
-			FSString name = homeUri.data();
-			Lsnprintf( buf, sizeof( buf ), "bad home path: %s\n", name.GetUtf8() );
-			NCMessageBox( g_MainWin, "Home", buf, true );
-		}
-		else
-		{
-			p->LoadPath( fs, path, 0, 0, PanelWin::SET );
-		}
-	}
-
-#else
-	const sys_char_t* home = ( sys_char_t* ) getenv( "HOME" );
-	if ( !home )
-	{
-		return;
-	}
-
-	FSPath path( sys_charset_id, home );
-	p->LoadPath( new FSSys(), path, 0, 0, PanelWin::SET );
-#endif
 }
