+++ src/drive-dlg.h
@@ -10,5 +10,3 @@
 class PanelWin;
 
 void SelectDriveDlg( PanelWin* p, PanelWin* OtherPanel );
-
-void OpenHomeDir( PanelWin* p, PanelWin* OtherPanel );
