+++ src/file-util.h
@@ -6,10 +6,14 @@
 
 #pragma once
 
+#include "globals.h"
+
+
 template <class T> class clPtr;
 class NCDialogParent;
 class FS;
 class FSPath;
+class PanelWin;
 
 
 /// Removes WCM Temp directory by the given Id
@@ -20,3 +24,13 @@ void RemoveAllWcmTempDirs();
 
 /// Loads to local temp file, returns local FS and Path back
 int LoadToTempFile( NCDialogParent* parent, clPtr<FS>* fs, FSPath* path );
+
+/// Opens HOME dir in the given panel
+void OpenHomeDir( PanelWin* p, PanelWin* OtherPanel );
+
+#ifdef _WIN32
+
+/// Returns HOME dir URL on Win32
+std::vector<unicode_t> GetHomeUriWin();
+
+#endif
